auth-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
sqlalchemy==2.0.28
sqlalchemy[asyncio]==2.0.28
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.3
pydantic-settings==2.2.1
python-multipart==0.0.9
email-validator==2.1.0
redis==5.0.1
